requirements.txt
# Flask-related dependencies
flask
flask-limiter
flask-login
flask-talisman
flask-wtf
wtforms

# Other dependencies
bcrypt
bleach
cryptography
gunicorn
itsdangerous
markdown
pyotp
python-dotenv
redis
qrcode[pil]
werkzeug